+++ gherkin_auto_complete_plus/commands/list_gherkin_steps_command.py
@@ -1,24 +1,22 @@
-import logging
-
 import sublime
 import sublime_plugin
 
-from .gherkin_event_listener import steps as catalogued_steps
-from .log_utilities import log_func
-from .settings import get_logging_level
+from ..gherkin_event_listener import steps as catalogued_steps
+from ..utilities import log_utilities, settings
 
-logging_level = get_logging_level()
+_logging_level = settings.get_logging_level()
+_logger = log_utilities.get_logger(__name__, _logging_level)
 
 
 class ListGherkinStepsCommand(sublime_plugin.WindowCommand):
     steps_dict = {}
 
-    @log_func(logging_level)
+    @log_utilities.log_function(_logging_level)
     def run(self):
         self.steps_dict = self.get_steps_dict(catalogued_steps)
         self.show_quick_panel(sorted(self.steps_dict.values()))
 
-    @log_func(logging_level)
+    @log_utilities.log_function(_logging_level)
     def get_steps_dict(self, catalogued_steps):
         steps = {}
         for index, step in enumerate(sorted(catalogued_steps)):
@@ -26,15 +24,19 @@ class ListGherkinStepsCommand(sublime_plugin.WindowCommand):
             steps[index] = formatted_step
         return steps
 
-    @log_func(logging_level)
+    @log_utilities.log_function(_logging_level)
     def on_done(self, index):
         if index == -1:
+            _logger.debug('index -1 (nothing selected from quick panel)')
             return
+
         target_step = self.steps_dict[index]
-        view = self.window.active_view()
+        _logger.debug('target step: {}'.format(target_step))
+
         mapping = {"characters": target_step}
+        view = self.window.active_view()
         view.run_command("insert", mapping)
 
-    @log_func(logging_level)
+    @log_utilities.log_function(_logging_level)
     def show_quick_panel(self, steps):
         self.window.show_quick_panel(steps, self.on_done)
